streamCTRL cleanup
