+++ src/CommandLineInterface/streamCTRL/streamCTRL_TUI.c
@@ -10,45 +10,16 @@
 
 #define _GNU_SOURCE
 
-/* =============================================================================================== */
-/* =============================================================================================== */
-/*                                        HEADER FILES                                             */
-/* =============================================================================================== */
-/* =============================================================================================== */
+
 
 #ifndef __STDC_LIB_EXT1__
 typedef int errno_t;
 #endif
 
-#include <malloc.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 #include <sys/file.h>
-#include <sys/mman.h> // mmap()
-
-#include <signal.h>
-#include <time.h>
-
-#include <sys/types.h>
-#include <unistd.h> // getpid() access()
-
 #include <sys/stat.h>
-#include <sys/types.h>
-
-#include <ctype.h>
-#include <fcntl.h>
 #include <ncurses.h>
-
-#include <dirent.h>
-
-#include <errno.h>
-#include <locale.h>
-#include <wchar.h>
-
 #include <pthread.h>
-
-#include <math.h>
 
 
 #include "CommandLineInterface/timeutils.h"
@@ -56,1282 +27,280 @@ typedef int errno_t;
 #include "CLIcore.h"
 #include "COREMOD_memory/COREMOD_memory.h"
 #include "COREMOD_tools/COREMOD_tools.h"
-#define SHAREDSHMDIR                                                           \
-    data.shmdir /**< default location of file mapped semaphores, can be over-ridden by env variable MILK_SHM_DIR */
+
+
+// default location of file mapped semaphores, can be over-ridden by env variable MILK_SHM_DIR
+#define SHAREDSHMDIR  data.shmdir
 
 #include "streamCTRL_TUI.h"
 
 #include "TUItools.h"
 
+#include "streamCTRL_find_streams.h"
+#include "streamCTRL_print_inode.h"
+#include "streamCTRL_print_procpid.h"
+#include "streamCTRL_print_trace.h"
+#include "streamCTRL_scan.h"
+#include "streamCTRL_utilfuncs.h"
+
 
 #define ctrl(x) ((x) &0x1f)
 
 
-
-/* =============================================================================================== */
-/* =============================================================================================== */
-/*                                      DEFINES, MACROS                                            */
-/* =============================================================================================== */
-/* =============================================================================================== */
-
-// shared memory access permission
-#define FILEMODE 0666
-
-#define STRINGLENMAX 32
-
-#define streamOpenNBpid_MAX 100
-#define nameNBchar          100
-#define PIDnameStringLen    12
-
-#define DISPLAY_MODE_HELP     1
-#define DISPLAY_MODE_SUMMARY  2
-#define DISPLAY_MODE_WRITE    3
-#define DISPLAY_MODE_READ     4
-#define DISPLAY_MODE_SPTRACE  5
-#define DISPLAY_MODE_FUSER    6
-
-#define PRINT_PID_DEFAULT          0
-#define PRINT_PID_FORCE_NOUPSTREAM 1
-
-#define NO_DOWNSTREAM_INDEX 100
-
-/* =============================================================================================== */
-/* =============================================================================================== */
-/*                                  GLOBAL DATA DECLARATION                                        */
-/* =============================================================================================== */
-/* =============================================================================================== */
-
 static short unsigned int wrow, wcol;
 
-/* =============================================================================================== */
-/* =============================================================================================== */
-/*                                    FUNCTIONS SOURCE CODE                                        */
-/* =============================================================================================== */
-/* =============================================================================================== */
 
-/**
- * @brief Returns ID number corresponding to a name
- *
- * @param images   pointer to array of images
- * @param name     input image name to be matched
- * @return imageID
- */
-imageID image_ID_from_images(IMAGE *images, const char *restrict name)
+
+// current streamCTRL TUI status
+
+struct streamCTRL_TUI_parameters
 {
-    imageID i;
+    int loopOK;
+    int dindexSelected;
+    int DisplayDetailLevel;
+    int DisplayMode;
+    int NBsindex;
+    int SORTING;
+    int DISPLAY_ALL_SEMS;
+    struct tm *uttime_lastScan;
+    int fuserScan;
+    int SORT_TOGGLE;
+    float frequ; // Hz
+    long ssindex[streamNBID_MAX]; // sorted index array
+} sTUIparam;
 
-    i = 0;
-    do
-    {
-        if(images[i].used == 1)
-        {
-            if((strncmp(name, images[i].name, strlen(name)) == 0) &&
-                    (images[i].name[strlen(name)] == '\0'))
-            {
-                clock_gettime(CLOCK_MILK, &images[i].md[0].lastaccesstime);
-                return i;
-            }
-        }
-        i++;
-    }
-    while(i != streamNBID_MAX);
 
-    return -1;
-}
 
-/**
- * @brief Returns first available ID in image array
- *
- * @param images     pointer to image array
- * @return imageID
- */
-imageID image_get_first_ID_available_from_images(IMAGE *images)
-{
-    imageID i;
 
-    i = 0;
-    do
-    {
-        if(images[i].used == 0)
-        {
-            images[i].used = 1;
-            return i;
-        }
-        i++;
-    }
-    while(i != streamNBID_MAX);
-    printf("ERROR: ran out of image IDs - cannot allocate new ID\n");
-    printf("NB_MAX_IMAGE should be increased above current value (%d)\n",
-           streamNBID_MAX);
 
-    return -1;
-}
 
-/**
- * @brief Get the process name by pid
- *
- * @param pid
- * @param pname
- * @return error code
- */
-errno_t get_process_name_by_pid(const int pid, char *pname)
-{
-    char *fname = (char *) calloc(STRINGMAXLEN_FULLFILENAME, sizeof(char));
 
-    WRITE_FULLFILENAME(fname, "/proc/%d/cmdline", pid);
-    FILE *fp = fopen(fname, "r");
-    if(fp)
-    {
-        size_t size;
-        size = fread(pname, sizeof(char), 1024, fp);
-        if(size > 0)
-        {
-            if('\n' == pname[size - 1])
-            {
-                pname[size - 1] = '\0';
-            }
-        }
-        fclose(fp);
-    }
 
-    free(fname);
-
-    return RETURN_SUCCESS;
-}
-
-/**
- * @brief Get the maximum PID value from system
- *
- * @return int
- */
-static int get_PIDmax()
-{
-    FILE *fp;
-    int   PIDmax;
-    int   fscanfcnt;
-
-    fp = fopen("/proc/sys/kernel/pid_max", "r");
-
-    fscanfcnt = fscanf(fp, "%d", &PIDmax);
-    if(fscanfcnt == EOF)
-    {
-        if(ferror(fp))
-        {
-            perror("fscanf");
-        }
-        else
-        {
-            fprintf(stderr,
-                    "Error: fscanf reached end of file, no matching "
-                    "characters, no matching failure\n");
-        }
-        exit(EXIT_FAILURE);
-    }
-    else if(fscanfcnt != 1)
-    {
-        fprintf(stderr,
-                "Error: fscanf successfully matched and assigned %i input "
-                "items, 1 expected\n",
-                fscanfcnt);
-        exit(EXIT_FAILURE);
-    }
-
-    fclose(fp);
-
-    return PIDmax;
-}
-
-struct streamCTRLarg_struct
-{
-    STREAMINFOPROC *streaminfoproc;
-    IMAGE          *images;
-};
-
-/** @brief find shared memory streams on system
- *
- * If filter is set to 1, require stream name to contain namefilter string
- * streaminfo needs to be pre-allocated
- *
- */
-
-int find_streams(STREAMINFO *streaminfo, int filter, const char *namefilter)
-{
-    int            NBstream = 0;
-    DIR           *d;
-    struct dirent *dir;
-
-    d = opendir(SHAREDSHMDIR);
-    if(d)
-    {
-        int sindex = 0;
-        while(((dir = readdir(d)) != NULL))
-        {
-            int   scanentryOK = 1;
-            char *pch         = strstr(dir->d_name, ".im.shm");
-
-            int matchOK = 0;
-
-            // name filtering (first pass, not exclusive to stream name, includes path and extension
-            if(filter == 1)
-            {
-                if(strstr(dir->d_name, namefilter) != NULL)
-                {
-                    matchOK = 1;
-                }
-            }
-            else
-            {
-                matchOK = 1;
-            }
-
-            if((pch) && (matchOK == 1))
-            {
-                // is file sym link ?
-                struct stat buf;
-                int         retv;
-                char        fullname[STRINGMAXLEN_FULLFILENAME];
-
-                WRITE_FULLFILENAME(fullname,
-                                   "%s/%s",
-                                   SHAREDSHMDIR,
-                                   dir->d_name);
-                retv = lstat(fullname, &buf);
-                if(retv == -1)
-                {
-                    endwin();
-                    printf("File \"%s\"", dir->d_name);
-                    perror("Error running lstat on file ");
-                    exit(EXIT_FAILURE);
-                }
-
-                if(S_ISLNK(buf.st_mode))  // resolve link name
-                {
-                    char  fullname[STRINGMAXLEN_FULLFILENAME];
-                    char *linknamefull;
-                    char  linkname[STRINGMAXLEN_FULLFILENAME];
-                    int   pathOK = 1;
-
-                    streaminfo[sindex].SymLink = 1;
-                    WRITE_FULLFILENAME(fullname,
-                                       "%s/%s",
-                                       SHAREDSHMDIR,
-                                       dir->d_name);
-                    linknamefull = realpath(fullname, NULL);
-
-                    if(linknamefull == NULL)
-                    {
-                        pathOK = 0;
-                    }
-                    else if(access(linknamefull,
-                                   R_OK)) // file cannot be read
-                    {
-                        pathOK = 0;
-                    }
-
-                    if(pathOK == 0)  // file cannot be read
-                    {
-                        scanentryOK = 0;
-                    }
-                    else
-                    {
-                        strcpy(linkname, basename(linknamefull));
-
-                        int          lOK = 1;
-                        unsigned int ii  = 0;
-                        while((lOK == 1) && (ii < strlen(linkname)))
-                        {
-                            if(linkname[ii] == '.')
-                            {
-                                linkname[ii] = '\0';
-                                lOK          = 0;
-                            }
-                            ii++;
-                        }
-                        strncpy(streaminfo[sindex].linkname,
-                                linkname,
-                                STRINGMAXLEN_STREAMINFO_NAME - 1);
-                    }
-
-                    if(linknamefull != NULL)
-                    {
-                        free(linknamefull);
-                    }
-                }
-                else
-                {
-                    streaminfo[sindex].SymLink = 0;
-                }
-
-                // get stream name
-                if(scanentryOK == 1)
-                {
-                    int strlencp1 = STRINGMAXLEN_STREAMINFO_NAME;
-                    int strlencp  = strlen(dir->d_name) - strlen(".im.shm");
-                    if(strlencp < strlencp1)
-                    {
-                        strlencp1 = strlencp;
-                    }
-                    strncpy(streaminfo[sindex].sname, dir->d_name, strlencp1);
-                    streaminfo[sindex]
-                    .sname[strlen(dir->d_name) - strlen(".im.shm")] = '\0';
-
-                    if(filter == 1)
-                    {
-                        if(strstr(streaminfo[sindex].sname, namefilter) !=
-                                NULL)
-                        {
-                            sindex++;
-                        }
-                    }
-                    else
-                    {
-                        sindex++;
-                    }
-                }
-            }
-        }
-
-        NBstream = sindex;
-    }
-    closedir(d);
-
-    return NBstream;
-}
-
-void *streamCTRL_scan(void *argptr)
-{
-    long NBsindex = 0;
-    long sindex   = 0;
-    long scancnt  = 0;
-
-    STREAMINFO *streaminfo;
-    char      **PIDname_array;
-
-    // timing
-    static int             firstIter = 1;
-    static struct timespec t0;
-    struct timespec        t1;
-    double                 tdiffv;
-    struct timespec        tdiff;
-
-    struct streamCTRLarg_struct *streamCTRLdata =
-        (struct streamCTRLarg_struct *) argptr;
-    STREAMINFOPROC *streaminfoproc = streamCTRLdata->streaminfoproc;
-    IMAGE          *images         = streamCTRLdata->images;
-
-    streaminfo    = streaminfoproc->sinfo;
-    PIDname_array = streaminfoproc->PIDtable;
-
-    streaminfoproc->loopcnt = 0;
-
-    // if set, write file list to file on first scan
-    //int WriteFlistToFile = 1;
-
-    FILE *fpfscan;
-
-    while(streaminfoproc->loop == 1)
-    {
-
-        // timing measurement
-        clock_gettime(CLOCK_MILK, &t1);
-        if(firstIter == 1)
-        {
-            tdiffv = 0.1;
-        }
-        else
-        {
-            tdiff  = timespec_diff(t0, t1);
-            tdiffv = 1.0 * tdiff.tv_sec + 1.0e-9 * tdiff.tv_nsec;
-        }
-        clock_gettime(CLOCK_MILK, &t0);
-        streaminfoproc->dtscan = tdiffv;
-
-        int mode = 0;
-        if(mode == 0)  // preferred mode
-        {
-            NBsindex = find_streams(streaminfo,
-                                    streaminfoproc->filter,
-                                    streaminfoproc->namefilter);
-
-            // write stream list to file if applicable
-            // ususally used for debugging only
-            //
-            if(streaminfoproc->WriteFlistToFile == 1)
-            {
-                fpfscan = fopen("streamCTRL_filescan.dat", "w");
-                fprintf(fpfscan, "# stream scan result\n");
-                fprintf(fpfscan,
-                        "filter: %d %s\n",
-                        streaminfoproc->filter,
-                        streaminfoproc->namefilter);
-                fprintf(fpfscan, "NBsindex = %ld\n", NBsindex);
-
-                for(sindex = 0; sindex < NBsindex; sindex++)
-                {
-                    //fprintf(fpfscan, "%4ld  %20s ", sindex, dir->d_name);
-
-                    if(streaminfo[sindex].SymLink == 1)
-                    {
-                        fprintf(fpfscan,
-                                "| %12s -> [ %12s ] ",
-                                streaminfo[sindex].sname,
-                                streaminfo[sindex].linkname);
-                    }
-                    else
-                    {
-                        fprintf(fpfscan,
-                                "| %12s -> [ %12s ] ",
-                                streaminfo[sindex].sname,
-                                " ");
-                    }
-                    fprintf(fpfscan, "\n");
-                }
-                fclose(fpfscan);
-            }
-
-            // Load into memory
-            for(sindex = 0; sindex < NBsindex; sindex++)
-            {
-                imageID ID;
-
-                ID = image_ID_from_images(images, streaminfo[sindex].sname);
-
-                // connect to stream
-                if(ID == -1)
-                {
-                    ID = image_get_first_ID_available_from_images(images);
-                    if(ID < 0)
-                    {
-                        return NULL;
-                    }
-                    ImageStreamIO_read_sharedmem_image_toIMAGE(
-                        streaminfo[sindex].sname,
-                        &images[ID]);
-                    streaminfo[sindex].deltacnt0          = 1;
-                    streaminfo[sindex].updatevalue        = 1.0;
-                    streaminfo[sindex].updatevalue_frozen = 1.0;
-                }
-                else
-                {
-                    float gainv = 1.0;
-                    if(firstIter == 0)
-                    {
-                        streaminfo[sindex].deltacnt0 =
-                            images[ID].md[0].cnt0 - streaminfo[sindex].cnt0;
-                        streaminfo[sindex].updatevalue =
-                            (1.0 - gainv) * streaminfo[sindex].updatevalue +
-                            gainv *
-                            (1.0 * streaminfo[sindex].deltacnt0 / tdiffv);
-                    }
-
-                    streaminfo[sindex].cnt0 =
-                        images[ID].md[0].cnt0; // keep memory of cnt0
-                    streaminfo[sindex].ID       = ID;
-                    streaminfo[sindex].datatype = images[ID].md[0].datatype;
-                }
-            }
-        }
-        else // candidate for removal
-        {
-            //---------------------------------------
-            DIR           *d;
-            struct dirent *dir;
-
-            d = opendir(SHAREDSHMDIR);
-            if(d)
-            {
-                sindex = 0;
-
-                while(((dir = readdir(d)) != NULL))
-                {
-                    int   scanentryOK = 1;
-                    char *pch         = strstr(dir->d_name, ".im.shm");
-
-                    int matchOK = 0;
-
-                    // name filtering
-                    if(streaminfoproc->filter == 1)
-                    {
-                        if(strstr(dir->d_name, streaminfoproc->namefilter) !=
-                                NULL)
-                        {
-                            matchOK = 1;
-                        }
-                    }
-                    else
-                    {
-                        matchOK = 1;
-                    }
-
-                    if((pch) && (matchOK == 1))
-                    {
-                        imageID ID;
-
-                        // is file sym link ?
-                        struct stat buf;
-                        int         retv;
-                        char        fullname[STRINGMAXLEN_FULLFILENAME];
-
-                        if(streaminfoproc->WriteFlistToFile == 1)
-                        {
-                            fprintf(fpfscan,
-                                    "%4ld  %20s ",
-                                    sindex,
-                                    dir->d_name);
-                        }
-
-                        WRITE_FULLFILENAME(fullname,
-                                           "%s/%s",
-                                           SHAREDSHMDIR,
-                                           dir->d_name);
-                        retv = lstat(fullname, &buf);
-                        if(retv == -1)
-                        {
-                            endwin();
-                            printf("File \"%s\"", dir->d_name);
-                            perror(
-                                "Error running lstat "
-                                "on file ");
-                            exit(EXIT_FAILURE);
-                        }
-
-                        if(S_ISLNK(buf.st_mode))  // resolve link name
-                        {
-                            char  fullname[STRINGMAXLEN_FULLFILENAME];
-                            char *linknamefull;
-                            char  linkname[STRINGMAXLEN_FULLFILENAME];
-                            int   pathOK = 1;
-
-                            streaminfo[sindex].SymLink = 1;
-                            WRITE_FULLFILENAME(fullname,
-                                               "%s/%s",
-                                               SHAREDSHMDIR,
-                                               dir->d_name);
-                            //                        readlink (fullname, linknamefull, 200-1);
-                            linknamefull = realpath(fullname, NULL);
-
-                            if(linknamefull == NULL)
-                            {
-                                pathOK = 0;
-                            }
-                            else if(access(linknamefull,
-                                           R_OK)) // file cannot be read
-                            {
-                                pathOK = 0;
-                            }
-
-                            if(pathOK == 0)  // file cannot be read
-                            {
-                                if(streaminfoproc->WriteFlistToFile == 1)
-                                {
-                                    fprintf(fpfscan,
-                                            " %s "
-                                            "<-> "
-                                            "LINK "
-                                            "%s "
-                                            "CANNOT"
-                                            " BE "
-                                            "READ "
-                                            "-> "
-                                            "off",
-                                            fullname,
-                                            linknamefull);
-                                }
-                                scanentryOK = 0;
-                            }
-                            else
-                            {
-                                strcpy(linkname, basename(linknamefull));
-
-                                int          lOK = 1;
-                                unsigned int ii  = 0;
-                                while((lOK == 1) && (ii < strlen(linkname)))
-                                {
-                                    if(linkname[ii] == '.')
-                                    {
-                                        linkname[ii] = '\0';
-                                        lOK          = 0;
-                                    }
-                                    ii++;
-                                }
-                                strncpy(streaminfo[sindex].linkname,
-                                        linkname,
-                                        STRINGMAXLEN_STREAMINFO_NAME - 1);
-                            }
-
-                            if(linknamefull != NULL)
-                            {
-                                free(linknamefull);
-                            }
-                        }
-                        else
-                        {
-                            streaminfo[sindex].SymLink = 0;
-                        }
-
-                        // get stream name and ID
-                        if(scanentryOK == 1)
-                        {
-                            int strlencp1 = STRINGMAXLEN_STREAMINFO_NAME;
-                            int strlencp =
-                                strlen(dir->d_name) - strlen(".im.shm");
-                            if(strlencp < strlencp1)
-                            {
-                                strlencp1 = strlencp;
-                            }
-                            strncpy(streaminfo[sindex].sname,
-                                    dir->d_name,
-                                    strlencp1);
-                            streaminfo[sindex].sname[strlen(dir->d_name) -
-                                                     strlen(".im.shm")] = '\0';
-
-                            if(streaminfoproc->WriteFlistToFile == 1)
-                            {
-                                if(streaminfo[sindex].SymLink == 1)
-                                {
-                                    fprintf(fpfscan,
-                                            "| "
-                                            "%12s "
-                                            "-> [ "
-                                            "%12s "
-                                            "] ",
-                                            streaminfo[sindex].sname,
-                                            streaminfo[sindex].linkname);
-                                }
-                                else
-                                {
-                                    fprintf(fpfscan,
-                                            "| "
-                                            "%12s "
-                                            "-> [ "
-                                            "%12s "
-                                            "] ",
-                                            streaminfo[sindex].sname,
-                                            " ");
-                                }
-                            }
-                        }
-
-                        if(scanentryOK == 1)
-                        {
-
-                            ID = image_ID_from_images(images,
-                                                      streaminfo[sindex].sname);
-
-                            // connect to stream
-                            if(ID == -1)
-                            {
-                                ID = image_get_first_ID_available_from_images(
-                                         images);
-                                if(ID < 0)
-                                {
-                                    return NULL;
-                                }
-                                ImageStreamIO_read_sharedmem_image_toIMAGE(
-                                    streaminfo[sindex].sname,
-                                    &images[ID]);
-                                streaminfo[sindex].deltacnt0          = 1;
-                                streaminfo[sindex].updatevalue        = 1.0;
-                                streaminfo[sindex].updatevalue_frozen = 1.0;
-                            }
-                            else
-                            {
-                                float gainv = 1.0;
-                                if(firstIter == 0)
-                                {
-                                    streaminfo[sindex].deltacnt0 =
-                                        images[ID].md[0].cnt0 -
-                                        streaminfo[sindex].cnt0;
-                                    streaminfo[sindex].updatevalue =
-                                        (1.0 - gainv) *
-                                        streaminfo[sindex].updatevalue +
-                                        gainv * (1.0 *
-                                                 streaminfo[sindex].deltacnt0 /
-                                                 tdiffv);
-                                }
-
-                                streaminfo[sindex].cnt0 =
-                                    images[ID]
-                                    .md[0]
-                                    .cnt0; // keep memory of cnt0
-                                streaminfo[sindex].ID = ID;
-                                streaminfo[sindex].datatype =
-                                    images[ID].md[0].datatype;
-
-                                sindex++;
-                            }
-                        }
-                        if(streaminfoproc->WriteFlistToFile == 1)
-                        {
-                            fprintf(fpfscan, "\n");
-                        }
-                    }
-                }
-
-                NBsindex = sindex;
-            }
-            closedir(d);
-
-            if(streaminfoproc->WriteFlistToFile == 1)
-            {
-                fclose(fpfscan);
-            }
-            //---------------------------------------
-        }
-
-        streaminfoproc->WriteFlistToFile = 0;
-
-        firstIter = 0;
-
-        if(streaminfoproc->fuserUpdate == 1)
-        {
-            FILE *fp;
-            int   STRINGMAXLEN_LINE = 2000;
-            char  plistoutline[STRINGMAXLEN_LINE];
-            char  command[STRINGMAXLEN_COMMAND];
-
-            int NBpid = 0;
-
-            //            sindexscan1 = ssindex[sindexscan];
-            int sindexscan1 = streaminfoproc->sindexscan;
-
-            if(streaminfoproc->sindexscan > NBsindex - 1)
-            {
-                streaminfoproc->fuserUpdate = 0;
-            }
-            else
-            {
-                int PReadMode = 1;
-
-                if(PReadMode == 0)
-                {
-                    // popen option
-                    {
-                        int slen = snprintf(command,
-                                            STRINGMAXLEN_COMMAND,
-                                            "/bin/fuser %s/%s.im.shm "
-                                            "2>/dev/null",
-                                            SHAREDSHMDIR,
-                                            streaminfo[sindexscan1].sname);
-                        if(slen < 1)
-                        {
-                            PRINT_ERROR("snprintf wrote <1 char");
-                            abort(); // can't handle this error any other way
-                        }
-                        if(slen >= STRINGMAXLEN_COMMAND)
-                        {
-                            PRINT_ERROR(
-                                "snprintf string "
-                                "truncation");
-                            abort(); // can't handle this error any other way
-                        }
-                    }
-
-                    fp = popen(command, "r");
-                    if(fp == NULL)
-                    {
-                        streaminfo[sindexscan1].streamOpenPID_status =
-                            2; // failed
-                    }
-                    else
-                    {
-                        streaminfo[sindexscan1].streamOpenPID_status = 1;
-
-                        if(fgets(plistoutline, STRINGMAXLEN_LINE - 1, fp) ==
-                                NULL)
-                        {
-                            snprintf(plistoutline, STRINGMAXLEN_LINE, " ");
-                        }
-                        pclose(fp);
-                    }
-                }
-                else
-                {
-                    // filesystem option
-                    char plistfname[STRINGMAXLEN_FULLFILENAME];
-                    WRITE_FULLFILENAME(plistfname,
-                                       "%s/%s.shmplist",
-                                       SHAREDSHMDIR,
-                                       streaminfo[sindexscan1].sname);
-
-                    {
-                        int slen = snprintf(command,
-                                            STRINGMAXLEN_COMMAND,
-                                            "/bin/fuser %s/%s.im.shm "
-                                            "2>/dev/null > %s",
-                                            SHAREDSHMDIR,
-                                            streaminfo[sindexscan1].sname,
-                                            plistfname);
-                        if(slen < 1)
-                        {
-                            PRINT_ERROR("snprintf wrote <1 char");
-                            abort(); // can't handle this error any other way
-                        }
-                        if(slen >= STRINGMAXLEN_COMMAND)
-                        {
-                            PRINT_ERROR(
-                                "snprintf string "
-                                "truncation");
-                            abort(); // can't handle this error any other way
-                        }
-                    }
-
-                    if(system(command) == -1)
-                    {
-                        perror("Command system() failed");
-                        exit(EXIT_FAILURE);
-                    }
-
-                    fp = fopen(plistfname, "r");
-                    if(fp == NULL)
-                    {
-                        streaminfo[sindexscan1].streamOpenPID_status = 2;
-                    }
-                    else
-                    {
-                        if(fgets(plistoutline, STRINGMAXLEN_LINE - 1, fp) ==
-                                NULL)
-                        {
-                            snprintf(plistoutline, STRINGMAXLEN_LINE, " ");
-                        }
-
-                        fclose(fp);
-                    }
-                }
-
-                if(streaminfo[sindexscan1].streamOpenPID_status != 2)
-                {
-                    char *pch;
-
-                    pch = strtok(plistoutline, " ");
-
-                    while(pch != NULL)
-                    {
-                        if(NBpid < streamOpenNBpid_MAX)
-                        {
-                            streaminfo[sindexscan1].streamOpenPID[NBpid] =
-                                atoi(pch);
-                            if(getpgid(streaminfo[sindexscan1]
-                                       .streamOpenPID[NBpid]) >= 0)
-                            {
-                                NBpid++;
-                            }
-                        }
-                        pch = strtok(NULL, " ");
-                    }
-                    streaminfo[sindexscan1].streamOpenPID_status = 1; // success
-                }
-
-                streaminfo[sindexscan1].streamOpenPID_cnt = NBpid;
-                // Get PID names
-                int pidIndex;
-                int cnt1 = 0;
-                for(pidIndex = 0;
-                        pidIndex < streaminfo[sindexscan1].streamOpenPID_cnt;
-                        pidIndex++)
-                {
-                    pid_t pid = streaminfo[sindexscan1].streamOpenPID[pidIndex];
-                    if((getpgid(pid) >= 0) && (pid != getpid()))
-                    {
-                        char *pname = (char *) calloc(1024, sizeof(char));
-                        get_process_name_by_pid(pid, pname);
-
-                        if(PIDname_array[pid] == NULL)
-                        {
-                            PIDname_array[pid] = (char *) malloc(
-                                                     sizeof(char) * (PIDnameStringLen + 1));
-                        }
-                        strncpy(PIDname_array[pid], pname, PIDnameStringLen);
-                        free(pname);
-                        cnt1++;
-                    }
-                }
-                streaminfo[sindexscan1].streamOpenPID_cnt1 = cnt1;
-
-                streaminfoproc->sindexscan++;
-            }
-        }
-
-        streaminfoproc->fuserUpdate0 = 0;
-
-        streaminfoproc->NBstream = NBsindex;
-        streaminfoproc->loopcnt++;
-
-        usleep(streaminfoproc->twaitus);
-
-        scancnt++;
-    }
-
-    return NULL;
-}
-
-static int streamCTRL_print_inode(ino_t  inode,
-                                  ino_t *upstreaminode,
-                                  int    NBupstreaminode,
-                                  int    downstreamindex)
-{
-    int Dispinode_NBchar = 9;
-    int is_upstream      = 0;
-    int is_downstream    = 0;
-    int upstreamindex    = 0;
-
-    for(int i = 0; i < NBupstreaminode; i++)
-    {
-        if(inode == upstreaminode[i])
-        {
-            is_upstream   = 1;
-            upstreamindex = i;
-            break;
-        }
-    }
-
-    if(downstreamindex < NO_DOWNSTREAM_INDEX)
-    {
-        is_downstream = 1;
-    }
-
-    if(is_upstream || is_downstream)
-    {
-        int colorcode = 3;
-        if(upstreamindex > 0)
-        {
-            colorcode = 7;
-        }
-
-        if(is_upstream)
-        {
-            //attron(COLOR_PAIR(colorcode));
-            screenprint_setcolor(colorcode);
-            TUI_printfw("%02d >", upstreamindex);
-            //attroff(COLOR_PAIR(colorcode));
-            screenprint_unsetcolor(colorcode);
-        }
-        else
-        {
-            TUI_printfw("    ");
-        }
-
-        TUI_printfw("-");
-
-        if(is_downstream)
-        {
-            int colorcode = 3;
-            if(downstreamindex > 0)
-            {
-                colorcode = 7;
-            }
-
-            //attron(COLOR_PAIR(colorcode));
-            screenprint_setcolor(colorcode);
-            TUI_printfw("> %02d", downstreamindex);
-            //attroff(COLOR_PAIR(colorcode));
-            screenprint_unsetcolor(colorcode);
-        }
-        else
-        {
-            TUI_printfw("    ");
-        }
-    }
-    else
-    {
-        TUI_printfw("%*d", Dispinode_NBchar, (int) inode);
-    }
-
-    return Dispinode_NBchar;
-}
-
-/** @brief print PID with highlighting
- *
- */
-static int streamCTRL_print_procpid(int      DispPID_NBchar,
-                                    pid_t    procpid,
-                                    pid_t   *upstreamproc,
-                                    int      NBupstreamproc,
-                                    uint32_t mode)
-{
-    //int DispPID_NBchar = 8;
-    int activitycolorcode = 0;
-    int is_upstream       = 0;
-    int upstreamindex     = 0;
-
-    if(mode & PRINT_PID_FORCE_NOUPSTREAM)
-    {
-        is_upstream = 0;
-    }
-    else
-    {
-        for(int i = 0; i < NBupstreamproc; i++)
-        {
-            if(procpid == upstreamproc[i])
-            {
-                is_upstream   = 1;
-                upstreamindex = i;
-                break;
-            }
-        }
-    }
-
-    if(procpid > 0)
-    {
-        if(getpgid(procpid) >= 0)  // check if pid active
-        {
-            activitycolorcode = 2;
-        }
-        else
-        {
-            if(procpid > 0)
-            {
-                activitycolorcode = 4;
-            }
-        }
-    }
-
-    if(is_upstream == 1)
-    {
-        if(activitycolorcode != 2)
-        {
-            screenprint_setreverse();
-            //attron(A_REVERSE);
-        }
-        else
-        {
-            activitycolorcode = 12;
-        }
-    }
-
-    if(activitycolorcode > 0)
-    {
-        screenprint_setcolor(activitycolorcode);
-        //attron(COLOR_PAIR(activitycolorcode));
-    }
-
-    if(is_upstream)
-    {
-        char upstreamstring[DispPID_NBchar + 1];
-        snprintf(upstreamstring, DispPID_NBchar, "%2d >>", upstreamindex);
-        TUI_printfw("%*s", DispPID_NBchar, upstreamstring);
-    }
-    else
-    {
-        TUI_printfw("%*d", DispPID_NBchar, (int) procpid);
-    }
-
-    if(activitycolorcode > 0)
-    {
-        screenprint_unsetcolor(activitycolorcode);
-        //attroff(COLOR_PAIR(activitycolorcode));
-    }
-
-    if((activitycolorcode != 2) && (is_upstream == 1))
-    {
-        screenprint_unsetreverse();
-        //attroff(A_REVERSE);
-    }
-
-    return DispPID_NBchar;
-}
-
-
-
-
-
-static errno_t streamCTRL_print_SPTRACE_details(
-    IMAGE   *streamCTRLimages,
-    imageID  ID,
-    pid_t   *upstreamproc,
-    int      NBupstreamproc,
-    uint32_t print_pid_mode
+static errno_t streamCTRL_keyinput_process(
+    int ch,
+    streamCTRLarg_struct *streamCTRLdata
 )
 {
-    /*
-       	int             triggermode;
-    	pid_t           procwrite_PID;
-    	ino_t           trigger_inode;
-    	struct timespec ts_procstart;
-    	struct timespec ts_streamupdate;
-    	int             trigsemindex;
-    	uint64_t        cnt0;
+    char c; // for user input
+    int  stringindex;
+    time_t  rawtime;
+    long sindex;
 
-    */
-
-    int Disp_inode_NBchar = 8;
-
-    int Disp_sname_NBchar = 16;
-
-    int Disp_cnt0_NBchar     = 12;
-    int Disp_PID_NBchar      = 8;
-    int Disp_type_NBchar     = 8;
-    int Disp_trigstat_NBchar = 12;
-
-    // suppress unused parameter warning
-    (void) print_pid_mode;
-
-    TUI_newline();
-    TUI_printfw("   %*s %*s %*s",
-                Disp_inode_NBchar,
-                "inode",
-                Disp_sname_NBchar,
-                "stream",
-                Disp_cnt0_NBchar,
-                "cnt0",
-                Disp_PID_NBchar,
-                "PID",
-                Disp_type_NBchar,
-                "type");
-    TUI_newline();
-
-    for(int spti = 0; spti < streamCTRLimages[ID].md[0].NBproctrace; spti++)
+    switch(ch)
     {
-        ino_t inode = streamCTRLimages[ID].streamproctrace[spti].trigger_inode;
-        int   sem   = streamCTRLimages[ID].streamproctrace[spti].trigsemindex;
-        pid_t pid   = streamCTRLimages[ID].streamproctrace[spti].procwrite_PID;
+    case 'x': // Exit control screen
+        sTUIparam.loopOK = 0;
+        break;
 
-        uint64_t cnt0 = streamCTRLimages[ID].streamproctrace[spti].cnt0;
-
-        TUI_printfw("%02d", spti);
-
-        TUI_printfw(" %*lu", Disp_inode_NBchar, inode);
-
-        // look for ID corresponding to inode
-        int IDscan  = 0;
-        int IDfound = -1;
-        while((IDfound == -1) && (IDscan < streamNBID_MAX))
+    case KEY_UP:
+        sTUIparam.dindexSelected--;
+        if(sTUIparam.dindexSelected < 0)
         {
-            if(streamCTRLimages[IDscan].used == 1)
-            {
-                if(streamCTRLimages[IDscan].md[0].inode == inode)
-                {
-                    IDfound = IDscan;
-                }
-            }
-            IDscan++;
+            sTUIparam.dindexSelected = 0;
         }
-        if(IDfound == -1)
+        break;
+
+    case KEY_DOWN:
+        sTUIparam.dindexSelected++;
+        if(sTUIparam.dindexSelected > sTUIparam.NBsindex - 1)
         {
-            TUI_printfw(" %*s", Disp_sname_NBchar, "???");
+            sTUIparam.dindexSelected = sTUIparam.NBsindex - 1;
+        }
+        break;
+
+    case KEY_PPAGE:
+        sTUIparam.dindexSelected -= 10;
+        if(sTUIparam.dindexSelected < 0)
+        {
+            sTUIparam.dindexSelected = 0;
+        }
+        break;
+
+    case KEY_LEFT:
+        sTUIparam.DisplayDetailLevel = 0;
+        break;
+
+    case KEY_RIGHT:
+        sTUIparam.DisplayDetailLevel = 1;
+        break;
+
+    case KEY_NPAGE:
+        sTUIparam.dindexSelected += 10;
+        if(sTUIparam.dindexSelected > sTUIparam.NBsindex - 1)
+        {
+            sTUIparam.dindexSelected = sTUIparam.NBsindex - 1;
+        }
+        break;
+
+    // ============ SCREENS
+
+    case 'h': // help
+        sTUIparam.DisplayMode = DISPLAY_MODE_HELP;
+        break;
+
+    case KEY_F(2): // semvals
+        sTUIparam.DisplayMode = DISPLAY_MODE_SUMMARY;
+        break;
+
+    case KEY_F(3): // write PIDs
+        sTUIparam.DisplayMode = DISPLAY_MODE_WRITE;
+        break;
+
+    case KEY_F(4): // read PIDs
+        sTUIparam.DisplayMode = DISPLAY_MODE_READ;
+        break;
+
+    case KEY_F(5): // read PIDs
+        sTUIparam.DisplayMode = DISPLAY_MODE_SPTRACE;
+        break;
+
+    case KEY_F(6): // open files
+        if((sTUIparam.DisplayMode == DISPLAY_MODE_FUSER) ||
+                (streamCTRLdata->streaminfoproc->fuserUpdate0 == 1))
+        {
+            streamCTRLdata->streaminfoproc->fuserUpdate = 1;
+            time(&rawtime);
+            sTUIparam.uttime_lastScan           = gmtime(&rawtime);
+            sTUIparam.fuserScan                 = 1;
+            streamCTRLdata->streaminfoproc->sindexscan = 0;
+        }
+
+        sTUIparam.DisplayMode = DISPLAY_MODE_FUSER;
+        //erase();
+        //TUI_printfw("SCANNING PROCESSES AND FILESYSTEM: PLEASE WAIT ...\n");
+        //refresh();
+        break;
+
+    // ============ ACTIONS
+
+    case ctrl('e'): // erase stream
+        sindex = sTUIparam.ssindex[sTUIparam.dindexSelected];
+        DEBUG_TRACEPOINT("removing stream sindex = %ld", sindex);
+        DEBUG_TRACEPOINT("removing stream ID = %ld", streaminfo[sindex].ID);
+
+        ImageStreamIO_destroyIm(&streamCTRLdata->images[streamCTRLdata->sinfo[sindex].ID]);
+
+        DEBUG_TRACEPOINT("%d", sTUIparam.dindexSelected);
+        break;
+
+    // ============ SCANNING
+
+    case '{': // slower scan update
+        streamCTRLdata->streaminfoproc->twaitus = (int)(1.2 * streamCTRLdata->streaminfoproc->twaitus);
+        if(streamCTRLdata->streaminfoproc->twaitus > 1000000)
+        {
+            streamCTRLdata->streaminfoproc->twaitus = 1000000;
+        }
+        break;
+
+    case '}': // faster scan update
+        streamCTRLdata->streaminfoproc->twaitus =
+            (int)(0.83333333333333333333 * streamCTRLdata->streaminfoproc->twaitus);
+        if(streamCTRLdata->streaminfoproc->twaitus < 1000)
+        {
+            streamCTRLdata->streaminfoproc->twaitus = 1000;
+        }
+        break;
+
+    case 'o': // output next scan to file
+        streamCTRLdata->streaminfoproc->WriteFlistToFile = 1;
+        break;
+
+    // ============ DISPLAY
+
+    case '-': // slower display update
+        sTUIparam.frequ *= 0.5;
+        if(sTUIparam.frequ < 1.0)
+        {
+            sTUIparam.frequ = 1.0;
+        }
+        if(sTUIparam.frequ > 64.0)
+        {
+            sTUIparam.frequ = 64.0;
+        }
+        break;
+
+    case '+': // faster display update
+        sTUIparam.frequ *= 2.0;
+        if(sTUIparam.frequ < 1.0)
+        {
+            sTUIparam.frequ = 1.0;
+        }
+        if(sTUIparam.frequ > 64.0)
+        {
+            sTUIparam.frequ = 64.0;
+        }
+        break;
+
+    case '1': // sorting by stream name
+        sTUIparam.SORTING = 1;
+        break;
+
+    case '2': // sorting by update freq (default)
+        sTUIparam.SORTING     = 2;
+        sTUIparam.SORT_TOGGLE = 1;
+        break;
+
+    case '3': // sort by number of processes accessing
+        sTUIparam.SORTING     = 3;
+        sTUIparam.SORT_TOGGLE = 1;
+        break;
+
+    case 'f': // stream name filter toggle
+        if(streamCTRLdata->streaminfoproc->filter == 0)
+        {
+            streamCTRLdata->streaminfoproc->filter = 1;
         }
         else
         {
-            TUI_printfw(" %*s",
-                        Disp_sname_NBchar,
-                        streamCTRLimages[IDfound].name);
+            streamCTRLdata->streaminfoproc->filter = 0;
         }
+        break;
 
-        TUI_printfw(" %*llu", Disp_cnt0_NBchar, cnt0);
-        TUI_printfw(" ");
-
-        Disp_PID_NBchar = streamCTRL_print_procpid(8,
-                          pid,
-                          upstreamproc,
-                          NBupstreamproc,
-                          PRINT_PID_FORCE_NOUPSTREAM);
-
-        TUI_printfw(" ");
-
-        switch(streamCTRLimages[ID].streamproctrace[spti].triggermode)
+    case 'F': // set stream name filter string
+        TUI_exit();
+        EXECUTE_SYSTEM_COMMAND("clear");
+        printf("Enter string: ");
+        fflush(stdout);
+        stringindex = 0;
+        while(((c = getchar()) != '\n') &&
+                (stringindex < STRINGLENMAX - 2))
         {
-            case PROCESSINFO_TRIGGERMODE_IMMEDIATE:
-                TUI_printfw("%d%*s",
-                            streamCTRLimages[ID].streamproctrace[spti].triggermode,
-                            Disp_type_NBchar - 1,
-                            "IMME");
-                break;
-
-            case PROCESSINFO_TRIGGERMODE_CNT0:
-                TUI_printfw("%d%*s",
-                            streamCTRLimages[ID].streamproctrace[spti].triggermode,
-                            Disp_type_NBchar - 1,
-                            "CNT0");
-                break;
-
-            case PROCESSINFO_TRIGGERMODE_CNT1:
-                TUI_printfw("%d%*s",
-                            streamCTRLimages[ID].streamproctrace[spti].triggermode,
-                            Disp_type_NBchar - 1,
-                            "CNT1");
-                break;
-
-            case PROCESSINFO_TRIGGERMODE_SEMAPHORE:
-                TUI_printfw("%d%*s",
-                            streamCTRLimages[ID].streamproctrace[spti].triggermode,
-                            Disp_type_NBchar - 4,
-                            "SM");
-                TUI_printfw(" %2d", sem);
-                break;
-
-            case PROCESSINFO_TRIGGERMODE_DELAY:
-                TUI_printfw("%d%*s",
-                            streamCTRLimages[ID].streamproctrace[spti].triggermode,
-                            Disp_type_NBchar - 1,
-                            "DELA");
-                break;
-
-            default:
-                TUI_printfw("%d%*s",
-                            streamCTRLimages[ID].streamproctrace[spti].triggermode,
-                            Disp_type_NBchar - 1,
-                            "UNKN");
-                break;
+            streamCTRLdata->streaminfoproc->namefilter[stringindex] = c;
+            if(c == 127)  // delete key
+            {
+                putchar(0x8);
+                putchar(' ');
+                putchar(0x8);
+                stringindex--;
+            }
+            else
+            {
+                //printf("[%d]", (int) c);
+                putchar(c); // echo on screen
+                stringindex++;
+            }
         }
-        TUI_printfw(" ");
+        printf("string entered\n");
+        streamCTRLdata->streaminfoproc->namefilter[stringindex] = '\0';
+        TUI_init_terminal(&wrow, &wcol);
+        break;
 
-        int print_timing = 0;
-        switch(streamCTRLimages[ID].streamproctrace[spti].triggerstatus)
-        {
-            case PROCESSINFO_TRIGGERSTATUS_WAITING:
-                TUI_printfw("%*s", Disp_trigstat_NBchar, "WAITING");
-                break;
-
-            case PROCESSINFO_TRIGGERSTATUS_RECEIVED:
-                screenprint_setcolor(2);
-                //attron(COLOR_PAIR(2));
-                TUI_printfw("%*s", Disp_trigstat_NBchar, "RECEIVED");
-                screenprint_unsetcolor(2);
-                //attroff(COLOR_PAIR(2));
-                print_timing = 1;
-                break;
-
-            case PROCESSINFO_TRIGGERSTATUS_TIMEDOUT:
-                //attron(COLOR_PAIR(3));
-                screenprint_setcolor(3);
-                TUI_printfw("%*s", Disp_trigstat_NBchar, "TIMEOUT");
-                //attroff(COLOR_PAIR(3));
-                screenprint_unsetcolor(3);
-                print_timing = 1;
-                break;
-
-            default:
-                TUI_printfw("%*s", Disp_trigstat_NBchar, "unknown");
-                break;
-        }
-
-        // trigger time
-        if(print_timing == 1)
-        {
-            TUI_printfw(
-                " at %ld.%09ld s",
-                streamCTRLimages[ID].streamproctrace[spti].ts_procstart.tv_sec,
-                streamCTRLimages[ID]
-                .streamproctrace[spti]
-                .ts_procstart.tv_nsec);
-
-            struct timespec tnow;
-            clock_gettime(CLOCK_MILK, &tnow);
-            struct timespec tdiff;
-
-            tdiff = timespec_diff(
-                        streamCTRLimages[ID].streamproctrace[spti].ts_procstart,
-                        tnow);
-            double tdiffv = 1.0 * tdiff.tv_sec + 1.0e-9 * tdiff.tv_nsec;
-
-            TUI_printfw("  %12.3f us ago", tdiffv * 1.0e6);
-        }
-
-        TUI_newline();
+    case 's': // toggle all sems / 2 sems
+        sTUIparam.DISPLAY_ALL_SEMS = !sTUIparam.DISPLAY_ALL_SEMS;
+        break;
     }
-
-    return RETURN_SUCCESS;
+    return EXIT_SUCCESS;
 }
+
+
+
+
+
+
+
+
+
 
 /**
  * @brief Control screen for stream structures
@@ -1341,24 +310,37 @@ static errno_t streamCTRL_print_SPTRACE_details(
 errno_t streamCTRL_CTRLscreen()
 {
 
+    // initialize sCTRLTUIparams
+    sTUIparam.loopOK = 1;
+    sTUIparam.dindexSelected = 0;
+    sTUIparam.DisplayDetailLevel = 0;
+    sTUIparam.DisplayMode      = DISPLAY_MODE_SUMMARY;
+    sTUIparam.NBsindex = 0;
+    sTUIparam.SORTING     = 0;
+    sTUIparam.DISPLAY_ALL_SEMS = 1; // Display all semaphores / just the first 2.
+    sTUIparam.fuserScan = 0;
+    sTUIparam.SORT_TOGGLE = 0;
+    sTUIparam.frequ = 32.0; // Hz
+
+
+
     int stringmaxlen = 300;
 
     // Display fields
     STREAMINFO    *streaminfo;
     STREAMINFOPROC streaminfoproc;
 
-    long sindex;           // scan index
-    long dindex;           // display index
+
+//    long dindex;           // display index
     long doffsetindex = 0; // offset index if more entries than can be displayed
 
-    long ssindex[streamNBID_MAX]; // sorted index array
 
-    float frequ = 32.0; // Hz
+
     int monstrlen = 200;
     char  monstring[monstrlen];
 
-    int SORTING     = 0;
-    int SORT_TOGGLE = 0;
+
+
 
     DEBUG_TRACEPOINT("function start ");
 
@@ -1390,7 +372,6 @@ errno_t streamCTRL_CTRLscreen()
     streaminfoproc.fuserUpdate      = 0;
 
     streaminfo = (STREAMINFO *) malloc(sizeof(STREAMINFO) * streamNBID_MAX);
-    streaminfoproc.sinfo = streaminfo;
     for(int sindex = 0; sindex < streamNBID_MAX; sindex++)
     {
         streaminfo[sindex].updatevalue          = 0.0;
@@ -1410,9 +391,11 @@ errno_t streamCTRL_CTRLscreen()
         streamCTRLimages[imID].semlog  = NULL;
     }
 
-    struct streamCTRLarg_struct streamCTRLdata;
+    streamCTRLarg_struct streamCTRLdata;
+    streamCTRLdata.sinfo          = streaminfo;
     streamCTRLdata.streaminfoproc = &streaminfoproc;
     streamCTRLdata.images         = streamCTRLimages;
+
 
     // catch signals (CTRL-C etc)
     //
@@ -1435,18 +418,11 @@ errno_t streamCTRL_CTRLscreen()
     DEBUG_TRACEPOINT("Initialize terminal");
     TUI_init_terminal(&wrow, &wcol);
 
-    int       NBsindex = 0;
-    int       loopOK   = 1;
+
     long long loopcnt  = 0;
 
-    int dindexSelected = 0;
 
-    int DisplayMode      = DISPLAY_MODE_SUMMARY;
-    int DISPLAY_ALL_SEMS = 1; // Display all semaphores / just the first 2.
 
-    struct tm *uttime_lastScan;
-    time_t     rawtime;
-    int        fuserScan = 0;
 
     streaminfoproc.filter       = 0;
     streaminfoproc.NBstream     = 0;
@@ -1495,8 +471,7 @@ errno_t streamCTRL_CTRLscreen()
 
     DEBUG_TRACEPOINT("Scan thread started");
 
-    char c; // for user input
-    int  stringindex;
+
 
     loopcnt = 0;
 
@@ -1505,25 +480,25 @@ errno_t streamCTRL_CTRLscreen()
     //        TUI_get_terminal_size(&wrow, &wcol);
 
     ino_t inodeselected      = 0;
-    int   DisplayDetailLevel = 0;
-    while(loopOK == 1)
+
+    while(sTUIparam.loopOK == 1)
     {
 
         int NBsinfodisp = wrow - 7;
 
         if(streaminfoproc.loopcnt == 1)
         {
-            SORTING     = 2;
-            SORT_TOGGLE = 1;
+            sTUIparam.SORTING     = 2;
+            sTUIparam.SORT_TOGGLE = 1;
         }
 
         //if(fuserUpdate != 1) // don't wait if ongoing fuser scan
 
-        usleep((long)(1000000.0 / frequ));
+        usleep((long)(1000000.0 / sTUIparam.frequ));
         //int ch = getch();
         int ch = get_singlechar_nonblock();
 
-        NBsindex = streaminfoproc.NBstream;
+        sTUIparam.NBsindex = streaminfoproc.NBstream;
 
         TUI_clearscreen(&wrow, &wcol);
 
@@ -1531,221 +506,20 @@ errno_t streamCTRL_CTRLscreen()
 
         DEBUG_TRACEPOINT("Process input character");
         //int selectedOK = 0; // goes to 1 if at least one process is selected
-        switch(ch)
-        {
-            case 'x': // Exit control screen
-                loopOK = 0;
-                break;
 
-            case KEY_UP:
-                dindexSelected--;
-                if(dindexSelected < 0)
-                {
-                    dindexSelected = 0;
-                }
-                break;
+        streamCTRL_keyinput_process(ch, &streamCTRLdata);
 
-            case KEY_DOWN:
-                dindexSelected++;
-                if(dindexSelected > NBsindex - 1)
-                {
-                    dindexSelected = NBsindex - 1;
-                }
-                break;
 
-            case KEY_PPAGE:
-                dindexSelected -= 10;
-                if(dindexSelected < 0)
-                {
-                    dindexSelected = 0;
-                }
-                break;
-
-            case KEY_LEFT:
-                DisplayDetailLevel = 0;
-                break;
-
-            case KEY_RIGHT:
-                DisplayDetailLevel = 1;
-                break;
-
-            case KEY_NPAGE:
-                dindexSelected += 10;
-                if(dindexSelected > NBsindex - 1)
-                {
-                    dindexSelected = NBsindex - 1;
-                }
-                break;
-
-            // ============ SCREENS
-
-            case 'h': // help
-                DisplayMode = DISPLAY_MODE_HELP;
-                break;
-
-            case KEY_F(2): // semvals
-                DisplayMode = DISPLAY_MODE_SUMMARY;
-                break;
-
-            case KEY_F(3): // write PIDs
-                DisplayMode = DISPLAY_MODE_WRITE;
-                break;
-
-            case KEY_F(4): // read PIDs
-                DisplayMode = DISPLAY_MODE_READ;
-                break;
-
-            case KEY_F(5): // read PIDs
-                DisplayMode = DISPLAY_MODE_SPTRACE;
-                break;
-
-            case KEY_F(6): // open files
-                if((DisplayMode == DISPLAY_MODE_FUSER) ||
-                        (streaminfoproc.fuserUpdate0 == 1))
-                {
-                    streaminfoproc.fuserUpdate = 1;
-                    time(&rawtime);
-                    uttime_lastScan           = gmtime(&rawtime);
-                    fuserScan                 = 1;
-                    streaminfoproc.sindexscan = 0;
-                }
-
-                DisplayMode = DISPLAY_MODE_FUSER;
-                //erase();
-                //TUI_printfw("SCANNING PROCESSES AND FILESYSTEM: PLEASE WAIT ...\n");
-                //refresh();
-                break;
-
-            // ============ ACTIONS
-
-            case ctrl('e'): // erase stream
-                DEBUG_TRACEPOINT("removing stream sindex = %ld", sindex);
-                sindex = ssindex[dindexSelected];
-                DEBUG_TRACEPOINT("removing stream sindex = %ld", sindex);
-                DEBUG_TRACEPOINT("removing stream ID = %ld", streaminfo[sindex].ID);
-
-                ImageStreamIO_destroyIm(&streamCTRLimages[streaminfo[sindex].ID]);
-
-                DEBUG_TRACEPOINT("%d", dindexSelected);
-                break;
-
-            // ============ SCANNING
-
-            case '{': // slower scan update
-                streaminfoproc.twaitus = (int)(1.2 * streaminfoproc.twaitus);
-                if(streaminfoproc.twaitus > 1000000)
-                {
-                    streaminfoproc.twaitus = 1000000;
-                }
-                break;
-
-            case '}': // faster scan update
-                streaminfoproc.twaitus =
-                    (int)(0.83333333333333333333 * streaminfoproc.twaitus);
-                if(streaminfoproc.twaitus < 1000)
-                {
-                    streaminfoproc.twaitus = 1000;
-                }
-                break;
-
-            case 'o': // output next scan to file
-                streaminfoproc.WriteFlistToFile = 1;
-                break;
-
-            // ============ DISPLAY
-
-            case '-': // slower display update
-                frequ *= 0.5;
-                if(frequ < 1.0)
-                {
-                    frequ = 1.0;
-                }
-                if(frequ > 64.0)
-                {
-                    frequ = 64.0;
-                }
-                break;
-
-            case '+': // faster display update
-                frequ *= 2.0;
-                if(frequ < 1.0)
-                {
-                    frequ = 1.0;
-                }
-                if(frequ > 64.0)
-                {
-                    frequ = 64.0;
-                }
-                break;
-
-            case '1': // sorting by stream name
-                SORTING = 1;
-                break;
-
-            case '2': // sorting by update freq (default)
-                SORTING     = 2;
-                SORT_TOGGLE = 1;
-                break;
-
-            case '3': // sort by number of processes accessing
-                SORTING     = 3;
-                SORT_TOGGLE = 1;
-                break;
-
-            case 'f': // stream name filter toggle
-                if(streaminfoproc.filter == 0)
-                {
-                    streaminfoproc.filter = 1;
-                }
-                else
-                {
-                    streaminfoproc.filter = 0;
-                }
-                break;
-
-            case 'F': // set stream name filter string
-                TUI_exit();
-                EXECUTE_SYSTEM_COMMAND("clear");
-                printf("Enter string: ");
-                fflush(stdout);
-                stringindex = 0;
-                while(((c = getchar()) != '\n') &&
-                        (stringindex < STRINGLENMAX - 2))
-                {
-                    streaminfoproc.namefilter[stringindex] = c;
-                    if(c == 127)  // delete key
-                    {
-                        putchar(0x8);
-                        putchar(' ');
-                        putchar(0x8);
-                        stringindex--;
-                    }
-                    else
-                    {
-                        //printf("[%d]", (int) c);
-                        putchar(c); // echo on screen
-                        stringindex++;
-                    }
-                }
-                printf("string entered\n");
-                streaminfoproc.namefilter[stringindex] = '\0';
-                TUI_init_terminal(&wrow, &wcol);
-                break;
-
-            case 's': // toggle all sems / 2 sems
-                DISPLAY_ALL_SEMS = !DISPLAY_ALL_SEMS;
-                break;
-        }
 
         DEBUG_TRACEPOINT("Input character processed");
 
-        if(dindexSelected < 0)
+        if(sTUIparam.dindexSelected < 0)
         {
-            dindexSelected = 0;
+            sTUIparam.dindexSelected = 0;
         }
-        if(dindexSelected > NBsindex - 1)
+        if(sTUIparam.dindexSelected > sTUIparam.NBsindex - 1)
         {
-            dindexSelected = NBsindex - 1;
+            sTUIparam.dindexSelected = sTUIparam.NBsindex - 1;
         }
 
         DEBUG_TRACEPOINT("Erase screen");
@@ -1768,7 +542,7 @@ errno_t streamCTRL_CTRLscreen()
 
         DEBUG_TRACEPOINT("Start display");
 
-        if(DisplayMode == DISPLAY_MODE_HELP)  // help
+        if(sTUIparam.DisplayMode == DISPLAY_MODE_HELP)  // help
         {
             //int attrval = A_BOLD;
 
@@ -1813,7 +587,7 @@ errno_t streamCTRL_CTRLscreen()
         else
         {
             DEBUG_TRACEPOINT(" ");
-            if(DisplayMode == DISPLAY_MODE_HELP)  // Inaccessible.
+            if(sTUIparam.DisplayMode == DISPLAY_MODE_HELP)  // Inaccessible.
             {
                 screenprint_setreverse();
                 TUI_printfw("[h] Help");
@@ -1825,7 +599,7 @@ errno_t streamCTRL_CTRLscreen()
             }
             TUI_printfw("   ");
 
-            if(DisplayMode == DISPLAY_MODE_SUMMARY)
+            if(sTUIparam.DisplayMode == DISPLAY_MODE_SUMMARY)
             {
                 screenprint_setreverse();
                 TUI_printfw("[F2] summary");
@@ -1837,7 +611,7 @@ errno_t streamCTRL_CTRLscreen()
             }
             TUI_printfw("   ");
 
-            if(DisplayMode == DISPLAY_MODE_WRITE)
+            if(sTUIparam.DisplayMode == DISPLAY_MODE_WRITE)
             {
                 screenprint_setreverse();
                 TUI_printfw("[F3] write PIDs");
@@ -1849,7 +623,7 @@ errno_t streamCTRL_CTRLscreen()
             }
             TUI_printfw("   ");
 
-            if(DisplayMode == DISPLAY_MODE_READ)
+            if(sTUIparam.DisplayMode == DISPLAY_MODE_READ)
             {
                 screenprint_setreverse();
                 TUI_printfw("[F4] read PIDs");
@@ -1861,7 +635,7 @@ errno_t streamCTRL_CTRLscreen()
             }
             TUI_printfw("   ");
 
-            if(DisplayMode == DISPLAY_MODE_SPTRACE)
+            if(sTUIparam.DisplayMode == DISPLAY_MODE_SPTRACE)
             {
                 screenprint_setreverse();
                 TUI_printfw("[F5] process traces");
@@ -1873,7 +647,7 @@ errno_t streamCTRL_CTRLscreen()
             }
             TUI_printfw("   ");
 
-            if(DisplayMode == DISPLAY_MODE_FUSER)
+            if(sTUIparam.DisplayMode == DISPLAY_MODE_FUSER)
             {
                 screenprint_setreverse();
                 TUI_printfw("[F6] access");
@@ -1890,7 +664,7 @@ errno_t streamCTRL_CTRLscreen()
                 "PIDmax = %d    Update frequ = %2d Hz  fscan=%5.2f Hz "
                 "( %5.2f Hz %5.2f %% busy ) ",
                 PIDmax,
-                (int)(frequ + 0.5),
+                (int)(sTUIparam.frequ + 0.5),
                 1.0 / streaminfoproc.dtscan,
                 1000000.0 / streaminfoproc.twaitus,
                 100.0 *
@@ -1903,21 +677,21 @@ errno_t streamCTRL_CTRLscreen()
                 screenprint_setcolor(9);
                 TUI_printfw("fuser scan ongoing  %4d  / %4d   ",
                             streaminfoproc.sindexscan,
-                            NBsindex);
+                            sTUIparam.NBsindex);
                 //attroff(COLOR_PAIR(9));
                 screenprint_unsetcolor(9);
             }
-            if(DisplayMode == DISPLAY_MODE_FUSER)
+            if(sTUIparam.DisplayMode == DISPLAY_MODE_FUSER)
             {
-                if(fuserScan == 1)
+                if(sTUIparam.fuserScan == 1)
                 {
                     TUI_printfw(
                         "Last scan on  %02d:%02d:%02d  - Press "
                         "F6 again to re-scan    C-c to stop "
                         "scan",
-                        uttime_lastScan->tm_hour,
-                        uttime_lastScan->tm_min,
-                        uttime_lastScan->tm_sec);
+                        sTUIparam.uttime_lastScan->tm_hour,
+                        sTUIparam.uttime_lastScan->tm_min,
+                        sTUIparam.uttime_lastScan->tm_sec);
                     TUI_newline();
                 }
                 else
@@ -1936,9 +710,9 @@ errno_t streamCTRL_CTRLscreen()
 
             int lastindex;
             lastindex = doffsetindex + NBsinfodisp;
-            if(lastindex > NBsindex - 1)
+            if(lastindex > sTUIparam.NBsindex - 1)
             {
-                lastindex = NBsindex - 1;
+                lastindex = sTUIparam.NBsindex - 1;
             }
 
             if(lastindex < 0)
@@ -1949,11 +723,11 @@ errno_t streamCTRL_CTRLscreen()
             TUI_printfw(
                 "%4d streams    Currently displaying %4d-%4d   "
                 "Selected %d  ID = %d  inode = %d",
-                NBsindex,
+                sTUIparam.NBsindex,
                 doffsetindex,
                 lastindex,
-                dindexSelected,
-                ssindex[dindexSelected],
+                sTUIparam.dindexSelected,
+                sTUIparam.ssindex[sTUIparam.dindexSelected],
                 (int) inodeselected);
 
             if(streaminfoproc.filter == 1)
@@ -1986,38 +760,38 @@ errno_t streamCTRL_CTRLscreen()
                         "   frequ ",
                         "#sem");
 
-            switch(DisplayMode)
+            switch(sTUIparam.DisplayMode)
             {
-                case DISPLAY_MODE_SUMMARY:
-                    TUI_printfw("     Semaphore values ....");
-                    TUI_newline();
-                    break;
+            case DISPLAY_MODE_SUMMARY:
+                TUI_printfw("     Semaphore values ....");
+                TUI_newline();
+                break;
 
-                case DISPLAY_MODE_WRITE:
-                    TUI_printfw("     write PIDs ....");
-                    TUI_newline();
-                    break;
+            case DISPLAY_MODE_WRITE:
+                TUI_printfw("     write PIDs ....");
+                TUI_newline();
+                break;
 
-                case DISPLAY_MODE_READ:
-                    TUI_printfw("     read PIDs ....");
-                    TUI_newline();
-                    break;
+            case DISPLAY_MODE_READ:
+                TUI_printfw("     read PIDs ....");
+                TUI_newline();
+                break;
 
-                case DISPLAY_MODE_SPTRACE:
-                    TUI_printfw(
-                        "     stream process traces:   \"(INODE "
-                        "TYPE/SEM PID)>\"");
-                    TUI_newline();
-                    break;
+            case DISPLAY_MODE_SPTRACE:
+                TUI_printfw(
+                    "     stream process traces:   \"(INODE "
+                    "TYPE/SEM PID)>\"");
+                TUI_newline();
+                break;
 
-                case DISPLAY_MODE_FUSER:
-                    TUI_printfw("     connected processes");
-                    TUI_newline();
-                    break;
+            case DISPLAY_MODE_FUSER:
+                TUI_printfw("     connected processes");
+                TUI_newline();
+                break;
 
-                default:
-                    TUI_newline();
-                    break;
+            default:
+                TUI_newline();
+                break;
             }
 
             screenprint_unsetbold();
@@ -2028,26 +802,25 @@ errno_t streamCTRL_CTRLscreen()
             // SORT
 
             // default : no sorting
-            for(dindex = 0; dindex < NBsindex; dindex++)
+            for(int dindex = 0; dindex < sTUIparam.NBsindex; dindex++)
             {
-                ssindex[dindex] = dindex;
+                sTUIparam.ssindex[dindex] = dindex;
             }
 
             DEBUG_TRACEPOINT(" ");
 
-            if(SORTING == 1)  // alphabetical sorting
+            if(sTUIparam.SORTING == 1)  // alphabetical sorting
             {
                 long *larray;
-                larray = (long *) malloc(sizeof(long) * NBsindex);
-                for(sindex = 0; sindex < NBsindex; sindex++)
+                larray = (long *) malloc(sizeof(long) * sTUIparam.NBsindex);
+                for(long sindex = 0; sindex < sTUIparam.NBsindex; sindex++)
                 {
                     larray[sindex] = sindex;
                 }
 
-                int sindex0, sindex1;
-                for(sindex0 = 0; sindex0 < NBsindex - 1; sindex0++)
+                for(int sindex0 = 0; sindex0 < sTUIparam.NBsindex - 1; sindex0++)
                 {
-                    for(sindex1 = sindex0 + 1; sindex1 < NBsindex; sindex1++)
+                    for(int sindex1 = sindex0 + 1; sindex1 < sTUIparam.NBsindex; sindex1++)
                     {
                         if(strcmp(streaminfo[larray[sindex0]].sname,
                                   streaminfo[larray[sindex1]].sname) > 0)
@@ -2059,57 +832,57 @@ errno_t streamCTRL_CTRLscreen()
                     }
                 }
 
-                for(dindex = 0; dindex < NBsindex; dindex++)
+                for(long dindex = 0; dindex < sTUIparam.NBsindex; dindex++)
                 {
-                    ssindex[dindex] = larray[dindex];
+                    sTUIparam.ssindex[dindex] = larray[dindex];
                 }
                 free(larray);
             }
 
             DEBUG_TRACEPOINT(" ");
 
-            if((SORTING == 2) ||
-                    (SORTING == 3)) // recent update and process access
+            if((sTUIparam.SORTING == 2) ||
+                    (sTUIparam.SORTING == 3)) // recent update and process access
             {
                 long   *larray;
                 double *varray;
-                larray = (long *) malloc(sizeof(long) * NBsindex);
-                varray = (double *) malloc(sizeof(double) * NBsindex);
+                larray = (long *) malloc(sizeof(long) * sTUIparam.NBsindex);
+                varray = (double *) malloc(sizeof(double) * sTUIparam.NBsindex);
 
-                if(SORT_TOGGLE == 1)
+                if(sTUIparam.SORT_TOGGLE == 1)
                 {
-                    for(sindex = 0; sindex < NBsindex; sindex++)
+                    for(long sindex = 0; sindex < sTUIparam.NBsindex; sindex++)
                     {
                         streaminfo[sindex].updatevalue_frozen =
                             streaminfo[sindex].updatevalue;
                     }
 
-                    if(SORTING == 3)
+                    if(sTUIparam.SORTING == 3)
                     {
-                        for(sindex = 0; sindex < NBsindex; sindex++)
+                        for(long sindex = 0; sindex < sTUIparam.NBsindex; sindex++)
                         {
                             streaminfo[sindex].updatevalue_frozen +=
                                 10000.0 * streaminfo[sindex].streamOpenPID_cnt1;
                         }
                     }
 
-                    SORT_TOGGLE = 0;
+                    sTUIparam.SORT_TOGGLE = 0;
                 }
 
-                for(sindex = 0; sindex < NBsindex; sindex++)
+                for(long sindex = 0; sindex < sTUIparam.NBsindex; sindex++)
                 {
                     larray[sindex] = sindex;
                     varray[sindex] = streaminfo[sindex].updatevalue_frozen;
                 }
 
-                if(NBsindex > 1)
+                if(sTUIparam.NBsindex > 1)
                 {
-                    quick_sort2l(varray, larray, NBsindex);
+                    quick_sort2l(varray, larray, sTUIparam.NBsindex);
                 }
 
-                for(dindex = 0; dindex < NBsindex; dindex++)
+                for(long dindex = 0; dindex < sTUIparam.NBsindex; dindex++)
                 {
-                    ssindex[NBsindex - dindex - 1] = larray[dindex];
+                    sTUIparam.ssindex[sTUIparam.NBsindex - dindex - 1] = larray[dindex];
                 }
 
                 free(larray);
@@ -2120,13 +893,13 @@ errno_t streamCTRL_CTRLscreen()
 
             // compute doffsetindex
 
-            while(dindexSelected - doffsetindex >
+            while(sTUIparam.dindexSelected - doffsetindex >
                     NBsinfodisp - 5) // scroll down
             {
                 doffsetindex++;
             }
 
-            while(dindexSelected - doffsetindex <
+            while(sTUIparam.dindexSelected - doffsetindex <
                     NBsinfodisp - 10) // scroll up
             {
                 doffsetindex--;
@@ -2142,18 +915,18 @@ errno_t streamCTRL_CTRLscreen()
             int DisplayFlag = 0;
 
             int print_pid_mode = PRINT_PID_DEFAULT;
-            for(dindex = 0; dindex < NBsindex; dindex++)
+            for(int dindex = 0; dindex < sTUIparam.NBsindex; dindex++)
             {
                 imageID ID;
-                sindex = ssindex[dindex];
+                int sindex = sTUIparam.ssindex[dindex];
                 ID     = streaminfo[sindex].ID;
 
                 while((streamCTRLimages[streaminfo[sindex].ID].used == 0) &&
-                        (dindex < NBsindex))
+                        (dindex < sTUIparam.NBsindex))
                 {
                     // skip this entry, as it is no longer in use
                     dindex++;
-                    sindex = ssindex[dindex];
+                    sindex = sTUIparam.ssindex[dindex];
                     ID     = streaminfo[sindex].ID;
                 }
 
@@ -2176,9 +949,9 @@ errno_t streamCTRL_CTRLscreen()
                     DisplayFlag = 0;
                 }
 
-                if(DisplayDetailLevel == 1)
+                if(sTUIparam.DisplayDetailLevel == 1)
                 {
-                    if(dindex == dindexSelected)
+                    if(dindex == sTUIparam.dindexSelected)
                     {
                         DisplayFlag = 1;
                     }
@@ -2190,7 +963,7 @@ errno_t streamCTRL_CTRLscreen()
 
                 DEBUG_TRACEPOINT(" ");
 
-                if(dindex == dindexSelected)
+                if(dindex == sTUIparam.dindexSelected)
                 {
                     DEBUG_TRACEPOINT(
                         "dindex %ld %d",
@@ -2293,7 +1066,7 @@ errno_t streamCTRL_CTRLscreen()
                     TUI_printfw(" ");
                 }
 
-                if((dindex == dindexSelected) && (DisplayDetailLevel == 0))
+                if((dindex == sTUIparam.dindexSelected) && (sTUIparam.DisplayDetailLevel == 0))
                 {
                     //attron(A_REVERSE);
                     screenprint_setreverse();
@@ -2344,7 +1117,7 @@ errno_t streamCTRL_CTRLscreen()
 
                 DEBUG_TRACEPOINT(" ");
 
-                if((DisplayMode < DISPLAY_MODE_FUSER) && (DisplayFlag == 1))
+                if((sTUIparam.DisplayMode < DISPLAY_MODE_FUSER) && (DisplayFlag == 1))
                 {
                     char str[STRINGMAXLEN_DEFAULT];
                     char str1[STRINGMAXLEN_DEFAULT];
@@ -2511,7 +1284,7 @@ errno_t streamCTRL_CTRLscreen()
 
                 if(streamCTRLimages[streaminfo[sindex].ID].md != NULL)
                 {
-                    if((DisplayMode == DISPLAY_MODE_SUMMARY) &&
+                    if((sTUIparam.DisplayMode == DISPLAY_MODE_SUMMARY) &&
                             (DisplayFlag == 1)) // sem vals
                     {
 
@@ -2522,7 +1295,7 @@ errno_t streamCTRL_CTRLscreen()
                         TUI_printfw(string);
 
                         int s;
-                        int max_s = DISPLAY_ALL_SEMS
+                        int max_s = sTUIparam.DISPLAY_ALL_SEMS
                                     ? streamCTRLimages[ID].md[0].sem
                                     : 3;
                         for(s = 0; s < max_s; s++)
@@ -2539,7 +1312,7 @@ errno_t streamCTRL_CTRLscreen()
                 DEBUG_TRACEPOINT(" ");
                 if(streamCTRLimages[streaminfo[sindex].ID].md != NULL)
                 {
-                    if((DisplayMode == DISPLAY_MODE_WRITE) &&
+                    if((sTUIparam.DisplayMode == DISPLAY_MODE_WRITE) &&
                             (DisplayFlag == 1)) // sem write PIDs
                     {
                         snprintf(string,
@@ -2550,7 +1323,7 @@ errno_t streamCTRL_CTRLscreen()
 
                         {
                             int s;
-                            int max_s = DISPLAY_ALL_SEMS
+                            int max_s = sTUIparam.DISPLAY_ALL_SEMS
                                         ? streamCTRLimages[ID].md[0].sem
                                         : 3;
                             for(s = 0; s < max_s; s++)
@@ -2565,7 +1338,7 @@ errno_t streamCTRL_CTRLscreen()
                             }
                         }
 
-                        if(DisplayDetailLevel == 1)
+                        if(sTUIparam.DisplayDetailLevel == 1)
                         {
 #ifdef IMAGESTRUCT_WRITEHISTORY
                             TUI_newline();
@@ -2658,7 +1431,7 @@ errno_t streamCTRL_CTRLscreen()
 
                 if(streamCTRLimages[streaminfo[sindex].ID].md != NULL)
                 {
-                    if((DisplayMode == DISPLAY_MODE_READ) &&
+                    if((sTUIparam.DisplayMode == DISPLAY_MODE_READ) &&
                             (DisplayFlag == 1)) // sem read PIDs
                     {
                         snprintf(string,
@@ -2668,7 +1441,7 @@ errno_t streamCTRL_CTRLscreen()
                         TUI_printfw(string);
 
                         int s;
-                        int max_s = DISPLAY_ALL_SEMS
+                        int max_s = sTUIparam.DISPLAY_ALL_SEMS
                                     ? streamCTRLimages[ID].md[0].sem
                                     : 3;
                         for(s = 0; s < max_s; s++)
@@ -2686,7 +1459,7 @@ errno_t streamCTRL_CTRLscreen()
 
                 if(streamCTRLimages[streaminfo[sindex].ID].md != NULL)
                 {
-                    if((DisplayMode == DISPLAY_MODE_SPTRACE) &&
+                    if((sTUIparam.DisplayMode == DISPLAY_MODE_SPTRACE) &&
                             (DisplayFlag == 1))
                     {
                         snprintf(string,
@@ -2713,29 +1486,29 @@ errno_t streamCTRL_CTRLscreen()
                                     .streamproctrace[spti]
                                     .triggermode)
                             {
-                                case PROCESSINFO_TRIGGERMODE_IMMEDIATE:
-                                    snprintf(string, stringlen, "(%7lu IM ", inode);
-                                    break;
+                            case PROCESSINFO_TRIGGERMODE_IMMEDIATE:
+                                snprintf(string, stringlen, "(%7lu IM ", inode);
+                                break;
 
-                                case PROCESSINFO_TRIGGERMODE_CNT0:
-                                    snprintf(string, stringlen, "(%7lu C0 ", inode);
-                                    break;
+                            case PROCESSINFO_TRIGGERMODE_CNT0:
+                                snprintf(string, stringlen, "(%7lu C0 ", inode);
+                                break;
 
-                                case PROCESSINFO_TRIGGERMODE_CNT1:
-                                    snprintf(string, stringlen, "(%7lu C1 ", inode);
-                                    break;
+                            case PROCESSINFO_TRIGGERMODE_CNT1:
+                                snprintf(string, stringlen, "(%7lu C1 ", inode);
+                                break;
 
-                                case PROCESSINFO_TRIGGERMODE_SEMAPHORE:
-                                    snprintf(string, stringlen, "(%7lu %02d ", inode, sem);
-                                    break;
+                            case PROCESSINFO_TRIGGERMODE_SEMAPHORE:
+                                snprintf(string, stringlen, "(%7lu %02d ", inode, sem);
+                                break;
 
-                                case PROCESSINFO_TRIGGERMODE_DELAY:
-                                    snprintf(string, stringlen, "(%7lu DL ", inode);
-                                    break;
+                            case PROCESSINFO_TRIGGERMODE_DELAY:
+                                snprintf(string, stringlen, "(%7lu DL ", inode);
+                                break;
 
-                                default:
-                                    snprintf(string, stringlen, "(%7lu ?? ", inode);
-                                    break;
+                            default:
+                                snprintf(string, stringlen, "(%7lu ?? ", inode);
+                                break;
                             }
                             TUI_printfw(string);
                             streamCTRL_print_procpid(8,
@@ -2746,7 +1519,7 @@ errno_t streamCTRL_CTRLscreen()
                             TUI_printfw(")> ");
                         }
 
-                        if(DisplayDetailLevel == 1)
+                        if(sTUIparam.DisplayDetailLevel == 1)
                         {
                             TUI_newline();
                             streamCTRL_print_SPTRACE_details(streamCTRLimages,
@@ -2756,10 +1529,10 @@ errno_t streamCTRL_CTRLscreen()
                                                              PRINT_PID_DEFAULT);
                         }
                     }
-                    if((DisplayMode == DISPLAY_MODE_SUMMARY) &&
+                    if((sTUIparam.DisplayMode == DISPLAY_MODE_SUMMARY) &&
                             (DisplayFlag == 1))
                     {
-                        if(DisplayDetailLevel == 1)
+                        if(sTUIparam.DisplayDetailLevel == 1)
                         {
                             TUI_newline();
                             TUI_newline();
@@ -2803,7 +1576,7 @@ errno_t streamCTRL_CTRLscreen()
                     }
                 }
 
-                if((DisplayMode == DISPLAY_MODE_FUSER) &&
+                if((sTUIparam.DisplayMode == DISPLAY_MODE_FUSER) &&
                         (DisplayFlag ==
                          1)) // list processes that are accessing streams
                 {
@@ -2820,45 +1593,45 @@ errno_t streamCTRL_CTRLscreen()
                     switch(streaminfo[sindex].streamOpenPID_status)
                     {
 
-                        case 1:
-                            streaminfo[sindex].streamOpenPID_cnt1 = 0;
-                            for(pidIndex = 0;
-                                    pidIndex < streaminfo[sindex].streamOpenPID_cnt;
-                                    pidIndex++)
+                    case 1:
+                        streaminfo[sindex].streamOpenPID_cnt1 = 0;
+                        for(pidIndex = 0;
+                                pidIndex < streaminfo[sindex].streamOpenPID_cnt;
+                                pidIndex++)
+                        {
+                            pid_t pid =
+                                streaminfo[sindex].streamOpenPID[pidIndex];
+                            streamCTRL_print_procpid(8,
+                                                     pid,
+                                                     upstreamproc,
+                                                     NBupstreamproc,
+                                                     print_pid_mode);
+
+                            if((getpgid(pid) >= 0) && (pid != getpid()))
                             {
-                                pid_t pid =
-                                    streaminfo[sindex].streamOpenPID[pidIndex];
-                                streamCTRL_print_procpid(8,
-                                                         pid,
-                                                         upstreamproc,
-                                                         NBupstreamproc,
-                                                         print_pid_mode);
 
-                                if((getpgid(pid) >= 0) && (pid != getpid()))
-                                {
+                                snprintf(string,
+                                         stringlen,
+                                         ":%-*.*s",
+                                         PIDnameStringLen,
+                                         PIDnameStringLen,
+                                         PIDname_array[pid]);
+                                TUI_printfw(string);
 
-                                    snprintf(string,
-                                             stringlen,
-                                             ":%-*.*s",
-                                             PIDnameStringLen,
-                                             PIDnameStringLen,
-                                             PIDname_array[pid]);
-                                    TUI_printfw(string);
-
-                                    streaminfo[sindex].streamOpenPID_cnt1++;
-                                }
+                                streaminfo[sindex].streamOpenPID_cnt1++;
                             }
-                            break;
+                        }
+                        break;
 
-                        case 2:
-                            snprintf(string, stringlen, "FAILED");
-                            TUI_printfw(string);
-                            break;
+                    case 2:
+                        snprintf(string, stringlen, "FAILED");
+                        TUI_printfw(string);
+                        break;
 
-                        default:
-                            snprintf(string, stringlen, "NOT SCANNED");
-                            TUI_printfw(string);
-                            break;
+                    default:
+                        snprintf(string, stringlen, "NOT SCANNED");
+                        TUI_printfw(string);
+                        break;
                     }
                 }
 
@@ -2866,7 +1639,7 @@ errno_t streamCTRL_CTRLscreen()
 
                 if(DisplayFlag == 1)
                 {
-                    if(dindex == dindexSelected)
+                    if(dindex == sTUIparam.dindexSelected)
                     {
                         screenprint_unsetreverse();
                         //attroff(A_REVERSE);
@@ -2907,7 +1680,7 @@ errno_t streamCTRL_CTRLscreen()
                 (data.signal_SEGV == 1) || (data.signal_HUP == 1) ||
                 (data.signal_PIPE == 1))
         {
-            loopOK = 0;
+            sTUIparam.loopOK = 0;
         }
 
         DEBUG_TRACEPOINT(" ");
