+++ src/CommandLineInterface/streamCTRL/streamCTRL_find_streams.h
@@ -1,0 +1,11 @@
+#ifndef _STREAMCTRL_FINDSTREAMS_H
+#define _STREAMCTRL_FINDSTREAMS_H
+
+
+int find_streams(
+    STREAMINFO *streaminfo,
+    int filter,
+    const char * __restrict namefilter
+);
+
+#endif
