+++ src/CommandLineInterface/streamCTRL/streamCTRL_scan.h
@@ -1,0 +1,9 @@
+#ifndef _STREAMCTRL_SCAN_H
+#define _STREAMCTRL_SCAN_H
+
+
+void *streamCTRL_scan(
+    void *argptr
+);
+
+#endif
